Double conversation fix2

+++ src/main/java/math/operation/ArithmeticOperations.java
@@ -19,6 +19,6 @@ public class ArithmeticOperations {
 	
 	public float divide(Integer a, Integer b)
 	{
-        return a/b;
+        return (float) a/b;
 	}
 }
